ci: Add GHA workflow to calculate the OpenSSF scorecard.
The [OpenSSF Scorecard](https://scorecard.dev/) is a common tool to
assess the quality of a dependency for "risky practices". While the
project itself calculates the score for a lot of projects, we would like
to add it directly here to ensure the "shortcomings" are visible to
maintainer and also because the workflow gets a bit more visibility into
the repository settings, it will also lead to a higher score.